modified

+++ src/SEARCHING/SquareRoot.java
@@ -1,10 +1,10 @@
 package SEARCHING;
 public class SquareRoot {
-    static int SquareRoot(int x) {
+    static float SquareRoot(int x) {
         int l=0;
         int r=x/2;
         int res=0;
-        int mid=0;
+        float mid=0;
         while (l<=r)
         {
              mid=(l+r)/2;
@@ -12,8 +12,8 @@ public class SquareRoot {
                 return mid;
          else if(mid*mid>x)
 
-             r=mid-1;
-         else l=mid+1;
+             r=(int)mid-1;
+         else l=(int)mid+1;
         }
         return mid;
     }
